move: `infrastructure` 以下のファイル構成を整理 * `*-dto.ts` を削除して、`*-response.ts` に統合 * Strapi で共有するファイルは `infrastructure/strapi` に移動
