+++ src/infrastructure/hint/hint-repository.ts
@@ -1,5 +1,5 @@
 import { Hint } from '@/domain/entities/hint';
-import { HintDTO } from '../dto/hint-dto';
+import { HintDTO } from './hint-response';
 
 export const newHintFromDTO = (hint: HintDTO, problemId: string) => {
   return new Hint(
