+++ src/infrastructure/language/lanuage-response.ts
@@ -2,9 +2,12 @@ import { z } from 'zod';
 import {
   strapiCommonAttributesDTO,
   strapiCommonDTO,
-} from './strapi-common-dto';
-import { ValidatorDTO, validatorDTO } from './validator-dto';
-import { InitialCodeDTO, initialCodeDTO } from './initial-code-dto';
+} from '../strapi/strapi-common-dto';
+import { ValidatorDTO, validatorDTO } from '../validator/validator-response';
+import {
+  InitialCodeDTO,
+  initialCodeDTO,
+} from '../initial-code/initial-code-response';
 import { supportedLanguage } from '@/domain/entities/supported-language';
 
 const baseLanguageAttributesDTO = strapiCommonAttributesDTO.extend({
