+++ src/infrastructure/problem/problem-repository.ts
@@ -1,8 +1,8 @@
 import { ProblemRepository } from '@/domain/repositories/problem-repository';
 import { Problem } from '@/domain/entities/problem';
 import { ok } from 'neverthrow';
-import { problemDTO, ProblemDTO } from '../dto/problem-dto';
-import { fetchStrapiData } from '../strapi-utils';
+import { problemDTO, ProblemDTO } from './problem-response';
+import { fetchStrapiData } from '../strapi/strapi-utils';
 
 export const newProblemFromDTO = (problem: ProblemDTO) => {
   return new Problem(
