+++ src/infrastructure/validator/validator-repository.ts
@@ -1,5 +1,5 @@
 import { Validator } from '@/domain/entities/validator';
-import { ValidatorDTO } from '../dto/validator-dto';
+import { ValidatorDTO } from './validator-response';
 import { SupportedLanguage } from '@/domain/entities/supported-language';
 
 export const newValidatorFromDTO = (
